MovieTile and MovieDetailsTile moved into `common/tiles` catalog. Move MissingPoster into a separate component.
