+++ src/common/tiles/MissingPoster/index.js
@@ -1,0 +1,7 @@
+import { Wrapper, StyledIconCamera } from "./styled";
+
+export const MissingPoster = () => (
+    <Wrapper>
+        <StyledIconCamera />
+    </Wrapper>
+);
