+++ src/common/tiles/MovieDetailsTile/styled.js
@@ -16,16 +16,25 @@ export const Wrapper = styled.article`
     }
 `;
 
-export const Poster = styled.img`
+export const PosterWrapper = styled.div`
     width: 312px;
     height: 464px;
-    border-radius: 5px;
     grid-row-end: span 2;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 
     @media (max-width: ${({ theme }) => theme.breakpoint.mobileMax}px) {
         width: 114px;
         height: 169px;
     }
+`;
+
+export const Poster = styled.img`
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 5px;  
 `;
 
 export const TileContent = styled.section`
