+++ src/common/tiles/MovieTile/Ratings/index.js
