+++ src/common/tiles/MovieTile/styled.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { ReactComponent as iconCamera } from "../../images/iconCamera.svg";
 import { Link } from "react-router-dom/cjs/react-router-dom.min";
 
 export const StyledLink = styled(Link)`
@@ -59,16 +58,6 @@ export const Poster = styled.img`
   }
 `;
 
-export const MissingPoster = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  border-radius: 5px;
-  background-color: ${({ theme }) => theme.color.postergrey};
-`;
-
 export const InfoContent = styled.div`
   display: flex;
   flex-direction: column;
@@ -105,19 +94,5 @@ export const ReleaseDate = styled.div`
   @media (max-width: ${({ theme }) => theme.breakpoint.mobileMax}px) {
     font-size: 13px;
     line-height: 1.3;
-  }
-`;
-
-export const StyledIconCamera = styled(iconCamera)`
-  width: 72px;
-  height: auto;
-  opacity: 50%;
-
-  path {
-    stroke-width: 1.69px;
-  }
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobileMax}px) {
-    width: 48px;
   }
 `;
