+++ src/features/Movies/MoviePage/index.js
@@ -2,7 +2,7 @@ import { useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
 import { useParams } from "react-router-dom/cjs/react-router-dom.min";
 import { useEffect } from "react";
-import { MovieDetailsTile } from "../../../common/MovieDetailsTile";
+import { MovieDetailsTile } from "../../../common/tiles/MovieDetailsTile";
 import { MainContent } from "../../../common/MainContent";
 import { Wrapper } from "../../../Wrapper";
 import { fetchMovieDetails, selectMovieDetails } from "../movieSlice";
